+++ NEWS.md
@@ -1,11 +1,13 @@
 # usmap 0.4.0.9999
 
 ### New Features
-* Ability to pass `ggplot2::geom_polygon()` parameters to `plot_usmap()`
-  * This is a breaking change and removes the `lines` parameter, as well as the built-in defaults. This may affect the appearance of your plots. The original defaults were `colour = lines, fill = "white", size = 0.4`.
+* Ability to pass `ggplot2::geom_polygon()` parameters to `plot_usmap()`, see [Issue #15](https://github.com/pdil/usmap/issues/15).
+  * This is a breaking change and removes the `lines` parameter. The previous defaults of `colour="black"`, `fill="white"`, and `size=0.4` are maintained and will be used for any of those parameters that are omitted.
   * Refer to the `ggplot2::geom_polygon()` documentation for more information.
-  * The following aesthetics are supported: `alpha`, `colour`, `fill`, `linetype`, `size`
+  * The following aesthetics are supported: `alpha`, `colour`/`color`, `fill`, `linetype`, `size`
   * This feature provides more direct control over the appearance of plots.
+* Ability to include county name labels on county maps, see [Issue #14](https://github.com/pdil/usmap/issues/14).
+  * They currently work the same as state labels except they include the full county name (including the word "County").
 
 
 # usmap 0.4.0